requirements.txt
langchain==0.2.17
langchain-google-genai==1.0.8
google-generativeai==0.7.2
pinecone-client==5.0.1
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.37.1
